requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.2.0
Django==4.2.3
djangorestframework==3.14.0
idna==3.4
marshmallow==3.17.0
packaging==23.1
Pillow==10.0.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
transbank-sdk==4.0.0
tzdata==2023.3
urllib3==2.0.3
django-crispy-forms
django-admin-interface